# db-assistant-ui/src/app/services/models/query-result-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface QueryResultDto {
  affectedRows?: number;
  aggregateQuery?: boolean;
  aggregateResult?: {
[key: string]: any;
};
  columns?: Array<string>;
  data?: Array<{
[key: string]: any;
}>;
  message?: string;
  queryType?: string;
  rowCount?: number;
  success?: boolean;
}