# db-assistant-ui/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenApi specification - marouane",
    "description": "OpenApi documentation for the db-assistant project with spring security",
    "contact": {
      "name": "marouane",
      "email": "marouaneyazza@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local env"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/sql/generate": {
      "post": {
        "tags": [
          "SQL Operations"
        ],
        "summary": "Generate SQL from a natural language question",
        "description": "Takes a plain-language question and converts it into an SQL query using the active database connection.",
        "operationId": "generateSql",
        "requestBody": {
          "description": "Natural language question to generate SQL",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "example": "Show all customers who spent more than 1000 last month"
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SqlResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/sql/execute": {
      "post": {
        "tags": [
          "SQL Operations"
        ],
        "summary": "Execute a raw SQL query",
        "description": "Runs the given SQL query on the active database connection and returns structured results.",
        "operationId": "executeSql",
        "requestBody": {
          "description": "SQL query string to execute on the active database connection",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SqlRequest"
              },
              "examples": {
                "Select example": {
                  "description": "Select example",
                  "value": {
                    "sql": "SELECT * FROM customers WHERE total_spent > 1000"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SqlResponseDto": {
        "type": "object",
        "properties": {
          "sql": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SqlRequest": {
        "type": "object",
        "properties": {
          "sql": {
            "type": "string"
          }
        }
      },
      "QueryResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rowCount": {
            "type": "integer",
            "format": "int32"
          },
          "queryType": {
            "type": "string"
          },
          "affectedRows": {
            "type": "integer",
            "format": "int32"
          },
          "aggregateResult": {
            "type": "object",
            "additionalProperties": {}
          },
          "aggregateQuery": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}